itra-gl-automation-functionapp-develop/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
azure-ai-formrecognizer==3.3.2
azure-common==1.1.28
azure-core==1.29.6
azure-functions==1.17.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-search-documents==11.4.0b11
azure-storage-blob==12.19.0
azure-storage-file-datalake==12.14.0
azure-storage-queue==12.9.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==41.0.7
distro==1.9.0
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
isodate==0.6.1
msal==1.26.0
msal-extensions==1.1.0
msrest==0.7.1
multidict==6.0.4
oauthlib==3.2.2
openai==1.6.1
packaging==23.2
pillow==10.2.0
portalocker==2.8.2
pycparser==2.21
pydantic==2.5.3
pydantic-core==2.14.6
PyJWT==2.8.0
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
pypdf==3.17.4
python-certifi-win32==1.6.1
python-dotenv==1.0.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
setuptools-scm==8.0.4
six==1.16.0
sniffio==1.3.0
tenacity==8.2.3
tiktoken==0.5.2
tomli==2.0.1
tqdm==4.66.1
typing-extensions==4.9.0
urllib3==2.1.0
wrapt==1.16.0
yarl==1.9.4
pyodbc==5.1.0
azure-data-tables==12.5.0